refactor: rename the components
